refactor build config and worker paths
