+++ posawesome/hooks.py
@@ -17,11 +17,8 @@ app_license = "GPLv3"
 # ------------------
 
 # include js, css files in header of desk.html
-app_include_css = ["posawesome.css"]
-# app_include_js = "/assets/posawesome/js/posawesome.js"
-app_include_js = [
-    "posawesome.bundle.js",
-]
+app_include_js  = ["posawesome.bundle.js"]
+app_include_css = ["js/posawesome.css"]
 
 # include js, css files in header of web template
 # web_include_css = "/assets/posawesome/css/posawesome.css"
