+++ posawesome/public/js/workers/itemWorker.js
